api/pyproject.toml
[tool.poetry]
name = "data-ingestion-backend"
version = "0.1.0"
description = ""
authors = ["Thinking Machines <hello@thinkingmachin.es>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
pydantic = {extras = ["email"], version = "^2.3.0"}
gunicorn = "^21.2.0"
pydantic-settings = "^2.0.3"
itsdangerous = "^2.1.2"
loguru = "^0.7.2"
fastapi-azure-auth = "^4.2.1"
python-multipart = "^0.0.6"
msal = "^1.24.1"
azure-storage-blob = "^12.18.3"
msgraph-sdk = "^1.0.0a16"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
ipython = "^8.16.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ["py311"]

[tool.isort]
profile = "black"
